--Finished the function to create cells to fill small gaps. Doesnt work as intended though. Perhaps change the way i store the tileMap data? As it is now the points of the edges of each 'pixel' are stored, not the center of the pixel, which is what i need to translate the location of the filler room to the correct location...

+++ DungeonGenerator/Room.h
@@ -1,5 +1,6 @@
 #pragma once
 #include <utility>
+#include <iostream>
 
 extern bool debug_flag;
 
@@ -44,8 +45,8 @@ private:
 	int vPosition[3];	//x,y, and z but z not used
 	int nHeight;
 	int nWidth;
-	float nHalfWidth = 0;
-	float nHalfHeight = 0;
+	int nHalfWidth = 0;
+	int nHalfHeight = 0;
 	float RoomSizeRatio;
 	bool bTrueRoom;
 	bool bFillerCell;
